feat: writing card design

+++ src/lib/utils.ts
@@ -3,23 +3,22 @@ import { twMerge } from "tailwind-merge";
 import { Post } from "#site/content";
 
 export function cn(...inputs: ClassValue[]) {
-  return twMerge(clsx(inputs));
+	return twMerge(clsx(inputs));
 }
 
 export function formateDate(input: string | number): string {
-  const date = new Date(input);
-  return date.toLocaleDateString("en-US", {
-    month: "long",
-    day: "numeric",
-    year: "numeric",
-  });
+	const date = new Date(input);
+	return date.toLocaleDateString("en-US", {
+		month: "numeric",
+		day: "numeric",
+		year: "numeric",
+	});
 }
 
 export function sortPosts(posts: Array<Post>) {
-  return posts.sort((a, b) => {
-    if (a.date > b.date) return -1;
-    if (a.date < b.date) return 1;
-    return 0;
-  });
+	return posts.sort((a, b) => {
+		if (a.date > b.date) return -1;
+		if (a.date < b.date) return 1;
+		return 0;
+	});
 }
-
